fix: linter cleanup and minor refactors in asset loading and shader uniform setting, move string helpers from config.cpp to a new file.

+++ src/render/UboBindings.h
@@ -8,7 +8,7 @@
 // Must stay in sync with `layout(std140, binding = N)` in GLSL shaders.
 namespace se::render {
 
-enum class UboBinding : uint32_t {
+enum class UboBinding : uint8_t {
     Frame = 0,
     Bones = 1,
     Terrain = 2,
